changes to Email feature
